+++ docs/conf.py
@@ -39,14 +39,14 @@ master_doc = "index"
 
 # General information about the project.
 project = "PyMuPDF"
-copyright = "2015-2019, Jorj X. McKie"
+copyright = "2015-2020, Jorj X. McKie"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = "1.16.11"
+release = "1.16.12"
 
 # The short X.Y version
 version = release
@@ -95,10 +95,10 @@ keep_warnings = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 # html_theme = "agogo"
-# html_theme = "sphinxdoc"
+html_theme = "sphinxdoc"
 # html_theme = "python_docs_theme"
 # html_theme = "sphinx_rtd_theme"
-html_theme = "classic"
+# html_theme = "classic"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -128,7 +128,7 @@ html_theme_options = {
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
-html_favicon = "PyMuPDF.ico"
+html_favicon = "pymupdf.ico"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,